Edits README to clarify email config must dos
The email_from value in config is set to email_from: `{"Your Name", "yourname@example.com"},` by default. If this is not changed emails might not go out as expected. Clarifying this in the README will help those that don't edit that value as they first start to wire things up with coherence.