Adds support for binary_ids
This commit adds support for binary ids in the migrations of the `coherence.install` mix task.

If the phoenix application is configured to default to binary ids, the generator of coherence will respect this when generating the migrations. This affects the `create` function as well as the `references` function in the migration.

The generated model does not need any special attributes, if the phoenix application is properly configured.