do not allow resetting the password with expired token
When resetting the password, the Coherence.PasswordController#edit/2
function checks, if the password reset token ist not expired. In case of
an expired reset token, the reset token gets removed.

This commit enables the same behaviour for the update/2 action, so
a user can not reset the password with an expired token by posting
the token and the password directly to the update action.