Added multilanguage support using Gettext
The translations use 'dgettext' with a "coherence" domain.
No pre-made translations are included in this commit, but they may be
added in the future.

It's impossible to test automatically whether the user interface
is completely internationalized. It can only be done by visual
inspection, so there might be holes.

Translations were added to:
- Coherence.ViewHelpers
- the templates (both email and forms)
- the *meta-templates* used to install coherence into a Phoenix
  application
- the email subject lines

A note about form labels:
-------------------------
Labels in forms now use explicit strings instead of atoms,
so that they can be internationalized by the user.
This was obviously not possible when 'Coherence.login.login_field'
is used, because it is not a string constant.
In those cases we can't do anything to help the user internationalize
that field. This is not too serious because those templates are
inserted into the user's Phoenix application, and as such they can be
easily customized.