Replaced CredentialStore Agent with a GenServer
Changes:
- Replaced Agent with GenServer
- We can now change the user data in all sessions and not only the current session.
  We can also change the data from outside a controller.
- Fixed some wrong @specs regarding credentials
- Added a 'credentials' type to enhance readability.

There are lot's of changes, but the commit makes sense as a whole
(except maybe for the tye definitions and fixing the @specs)

Now, in more detail:

*Replaced Agent with a GenServer*
The API remains backward compatible, except for the fact that all
references to Agent are now references to a Server.

A GenServer is a better choice because it abstracts the state further
and allows for more complex queries of the session data.

In this case, a GenServer was necessary to have a fast way of getting
all sessions belonging to a certain user without iterating over all
sessions.

Having a GenServer allows us to keep an (opaque) index, hidden from the
end user (as it should be, given it's an implementation detail).
This extra decoupling between querying the state and the way the state
is in fact stores looks like a good thing.

This GenServer is implemented using ExActor.
ExActor gives an even nicer API on top of the raw GenServer,
and reduces boilerplate by a lot.

It does introduce an extra dependency, but it's compile-time only,
as all macros are expanded at compile-time into a normal GenServer.
The mapping between an ExActor and a GenServer is straightforward,
so I think it doesn't loose much in clarity.
In fact, I think redability is enhanced by not using a raw GenServer.

*Changing the user's data in all sessions*

It requires keeping an additional index in the Credential store,
but it's worth it because it makes it very easy to change the user's
data. Previously, the user had to log out and login so that changes
would be reflected in all sessions.