Fix issue when attempting to immediately log in after registering an account
When attempting to log in an error would occur as the Html Responder session_create_success expects notice in it's opts and it is never provided.

Added a new "account_created_successfully" message and provided this with the required opts when attempting to log in immediately after creating an account.